new function, error with -int

+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: ybuhai <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/11 11:37:00 by ybuhai            #+#    #+#             */
-/*   Updated: 2018/12/18 13:19:41 by ybuhai           ###   ########.fr       */
+/*   Updated: 2018/12/21 17:59:48 by ybuhai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ typedef struct		g_stuct
 g_data		g_flags;
 int		ft_printf(const char *format, ...);
 void	ft_putchar(int c);
-int		intlen(int n);
+int		intlen(long long int n);
 int		check_width(const char *str);
 int		check_length(const char *str);
 int		check_symbol(char c);
@@ -37,7 +37,7 @@ int		check_precision(const char *str);
 void	prepare_to_print(va_list argptr);
 void	print_char(int c);
 void	print_str(char *str, int i);
-void	print_int(int nbr);
+void	print_int(long long int nbr);
 void	struct_to_nul(void);
 
 #endif
